+++ vscode_extension/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Version history
 ## 0.3.24
-- Disable the `Copy Symbol to Clipboard` command when there is a text selection.
+- `Copy Symbol to Clipboard`
+  - Disable command when there is a text selection.
+  - Show a progress dialog when Sorbet is not ready to process commands.
 
 ## 0.3.23
 - Fix: Sorbet extension fails when opening a project containing Ruby code but no active configuration.